requirements.txt
tensorflow
flask
keras
numpy
pandas
matplotlib
scikit-learn
Pillow
gunicornadd